Basically functional.
Can connect to api and send a file, but fails because it doesn't take account of the DataAdded acks and overwhelms the server.
Console.WriteLine's and debug statements are a bit random and I need to check whether the async code is as simple as it could be. I would have thought that an AutoResetEvent wasn't necessary and there should be a way to do Task.Wait.